Extracted rest of destructurers tests to files.
